Updating install point for jupyter-user, moving codeserver into jupyter-user image (#505)
* Updating install point for jupyter-user, moving codeserver out of apps into jupyter-user image

* static checks...NICE

* Static Checks Deux

* Static Check - Thrice....and I give in...
