+++ cli/aws_orbit/data/charts/team/jupyter-hub/src/jupyterhub_config.py
@@ -66,7 +66,7 @@ c.KubeSpawner.pvc_name_template = pvc_name_template
 c.KubeSpawner.volumes = [{"name": "efs-volume", "persistentVolumeClaim": {"claimName": "jupyterhub"}}]
 c.KubeSpawner.volume_mounts = [{"mountPath": "/efs", "name": "efs-volume"}]
 c.KubeSpawner.fs_gid = 100
-c.KubeSpawner.lifecycle_hooks = {"postStart": {"exec": {"command": ["/bin/sh", "/home/jovyan/.orbit/bootstrap.sh"]}}}
+c.KubeSpawner.lifecycle_hooks = {"postStart": {"exec": {"command": ["/bin/sh", "/opt/orbit/bootstrap.sh"]}}}
 c.KubeSpawner.node_selector = {"orbit/usage": "teams", "orbit/node-type": "ec2"}
 c.KubeSpawner.service_account = f"{TEAM}"
 c.JupyterHub.allow_named_servers = True
